refactor : adjusting application layers
