+++ src/application/errors/app-error.ts
