+++ src/application/usecases/multi-upload-file-service.ts
@@ -1,7 +1,7 @@
 import { S3Client, PutObjectCommand } from '@aws-sdk/client-s3'
 import { Logger } from 'winston'
-import { AppError, FileLengthError } from '../utils/errors/app-error'
-import { HttpStatusCode } from '../utils/http-status'
+import { FileLengthError, AppError } from '../errors/app-error'
+import { HttpStatusCode } from '../../domain/http-status'
 
 export class UploadFilesService {
   private s3: S3Client
@@ -39,7 +39,7 @@ export class UploadFilesService {
         )
       }
     })
-   
+
     await Promise.all(FilesToUpload)
   }
 }
