+++ src/application/usecases/upload-file-service.ts
@@ -4,8 +4,8 @@ import {
   S3Client,
 } from '@aws-sdk/client-s3'
 import { Logger } from 'winston'
-import { AppError, FileNotFound } from '../utils/errors/app-error'
-import { HttpStatusCode } from '../utils/http-status'
+import { AppError, FileNotFound } from '../errors/app-error'
+import { HttpStatusCode } from '../../domain/http-status'
 
 export class UploadFileService {
   private s3: S3Client
