+++ src/infra/aws/index.ts
@@ -1,5 +1,5 @@
 import { S3Client } from '@aws-sdk/client-s3'
-import { env } from '../config/env'
+import { env } from '../infra/config/env'
 
 export const s3 = new S3Client({
   credentials: {
